requirements.txt
torch==1.9.0
torchvision==0.10.0
pytorch_lightning==1.4.2
transformers==4.9.2
datasets==1.11.0
scikit-image==0.18.2
Pillow==8.3.1
matplotlib==3.4.2
